Subjects/ПМС/Examples/usart-comport-telegram-experiments-master/web/keyboard/requirements.txt
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.0.2
future==0.16.0
gevent==1.3.6
gevent-websocket==0.10.1
greenlet==0.4.15
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyserial==3.4
python-engineio==2.3.2
python-socketio==2.0.0
PyYAML==3.13
six==1.11.0
Werkzeug==0.14.1